Add concurrency task for fetching images
