+++ app/models/profile.js
@@ -1,11 +1,15 @@
 import DS from 'ember-data';
 import { attr } from '@ember-decorators/data';
+import convert from 'xml-js';
+import { isArray } from '@ember/array';
+import { computed } from '@ember-decorators/object';
+import { restartableTask } from 'ember-concurrency-decorators';
 
 const { Model } = DS;
 
-export default class ProfileModel extends Model {
-  @attr() images;
+const S3_BUCKET_HOST = 'https://waterfront-access-photos.nyc3.digitaloceanspaces.com';
 
+export default class ProfileModel extends Model {
   @attr('string') park_name;
   @attr('string') paws_id;
   @attr('string') park_name;
@@ -45,4 +49,23 @@ export default class ProfileModel extends Model {
   @attr('boolean') shore_public_walkway;
   @attr('boolean') trash_receptacles;
   @attr('boolean') trees;
+
+  @restartableTask()
+  images = function* () { // eslint-disable-line
+    const id = this.get('paws_id');
+
+    return fetch(`${S3_BUCKET_HOST}/?prefix=${id}`)
+      .then(d => d.text())
+      .then(d => {
+        const json = convert.xml2js(d, { compact: true })
+        let { Contents: contents = [] } = json.ListBucketResult;
+
+        const contentsArray = isArray(contents) ? contents : [contents];
+
+        return contentsArray.map(d => {
+          const filename = d.Key._text;
+          return `${S3_BUCKET_HOST}/${filename}`;
+        });
+      });
+  }
 }
