+++ app/routes/application.js
@@ -8,7 +8,7 @@ export default class ApplicationRoute extends Route {
     const layerGroups = await this.store.query('layer-group', {
       'layer-groups': [
         { id: 'subway', visible: true },
-        { id: 'waterfront-access--publically-owned', visible: true },
+        { id: 'waterfront-access--publicly-owned', visible: true },
         { id: 'waterfront-access--wpaas', visible: true },
       ],
     });
