+++ package.json
@@ -43,6 +43,8 @@
     "ember-cli-slick": "^2.0.0",
     "ember-cli-sri": "^2.1.0",
     "ember-cli-uglify": "^2.0.0",
+    "ember-concurrency": "^0.8.21",
+    "ember-concurrency-decorators": "^0.0.4",
     "ember-data": "~3.3.0",
     "ember-decorators": "^2.3.1",
     "ember-export-application-global": "^2.0.0",
